Sign In page, Button and Input components
Added some features like on focus change and on blur change to the input element and managed all data that is being putted in input fields in one json object. It is possible just with unform that compiles all that data. Beyond that added the sign up background for further utilization. Also input auto manage itself with some ref datas and it prints errors catched.