A lot of changes, details on description
A lot of changes to the project like, auth context that is a context in wich the application gets the token and the user who is logged, an toast that notify if a login was failed or an sign up was done and some error handling for types verification with Yup at registration and login with some throw error that gives the abillity to show the toast messages. Also added a dashboard that is a simple page where the person who's logged is redirected and some verification if it's trully logged.